feat(cli): add installable package and unified command
