+++ .github/workflows/ci.yml
@@ -17,6 +17,8 @@ concurrency:
 jobs:
   core-tests:
     name: Core tests (${{ matrix.os }}, Python ${{ matrix.python }})
+    env:
+      PYTHONPATH: ${{ github.workspace }}/src
     strategy:
       fail-fast: false
       matrix:
@@ -43,9 +45,13 @@ jobs:
         run: python -m unittest discover -s tests -v
       - name: Run Draw.io QA regression tests
         run: python scripts/test_drawio_qa.py -v
+      - name: Test packaged install outside the repository
+        run: python scripts/package_smoke.py
 
   skill-validation:
     name: Skill and repository validation
+    env:
+      PYTHONPATH: ${{ github.workspace }}/src
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v6
@@ -53,20 +59,22 @@ jobs:
         with:
           python-version: "3.12"
       - run: python scripts/validate_skill.py
-      - run: python scripts/doctor.py
+      - run: python scripts/nexcanvas_cli.py doctor
 
   reference-validation:
     name: Reference project postflight
+    env:
+      PYTHONPATH: ${{ github.workspace }}/src
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v6
       - uses: actions/setup-python@v6
         with:
           python-version: "3.12"
-      - run: python scripts/postflight.py examples/v2-rag-reference
-      - run: python scripts/postflight.py examples/v3-dense-industrial-ai
-      - run: python scripts/postflight.py examples/v4-hub-spoke-industrial
-      - run: python scripts/postflight.py examples/v5-multi-agent-workflow
+      - run: python scripts/nexcanvas_cli.py postflight examples/v2-rag-reference
+      - run: python scripts/nexcanvas_cli.py postflight examples/v3-dense-industrial-ai
+      - run: python scripts/nexcanvas_cli.py postflight examples/v4-hub-spoke-industrial
+      - run: python scripts/nexcanvas_cli.py postflight examples/v5-multi-agent-workflow
 
   quality-gate:
     name: Quality gate
