+++ .github/workflows/release-please.yml
@@ -28,6 +28,18 @@ jobs:
         with:
           ref: ${{ steps.release.outputs.tag_name }}
 
+      - name: Set up release Python
+        if: ${{ steps.release.outputs.release_created }}
+        uses: actions/setup-python@v6
+        with:
+          python-version: "3.12"
+
+      - name: Build Python wheel and source distribution
+        if: ${{ steps.release.outputs.release_created }}
+        run: |
+          python -m pip install --upgrade build
+          python -m build
+
       - name: Build checksummed skill bundle
         if: ${{ steps.release.outputs.release_created }}
         env:
@@ -35,11 +47,11 @@ jobs:
         run: |
           archive="nexcanvas-drawio-${RELEASE_TAG}.zip"
           git archive --format=zip --output="$archive" "$RELEASE_TAG"
-          sha256sum "$archive" > SHA256SUMS.txt
+          sha256sum "$archive" dist/* > SHA256SUMS.txt
 
       - name: Upload release assets
         if: ${{ steps.release.outputs.release_created }}
         env:
           GH_TOKEN: ${{ github.token }}
           RELEASE_TAG: ${{ steps.release.outputs.tag_name }}
-        run: gh release upload "$RELEASE_TAG" "nexcanvas-drawio-${RELEASE_TAG}.zip" SHA256SUMS.txt --clobber
+        run: gh release upload "$RELEASE_TAG" "nexcanvas-drawio-${RELEASE_TAG}.zip" dist/* SHA256SUMS.txt --clobber
