+++ pyproject.toml
@@ -1,0 +1,49 @@
+[build-system]
+requires = ["setuptools>=68"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "nexcanvas-drawio"
+dynamic = ["version"]
+description = "Portable agent skill and deterministic toolkit for editable, reference-grade Draw.io architecture diagrams."
+readme = "README.md"
+requires-python = ">=3.10"
+license = { file = "LICENSE" }
+authors = [{ name = "NexCanvas contributors" }]
+keywords = ["drawio", "architecture", "diagramming", "agent-skill", "mxgraph"]
+classifiers = [
+  "Development Status :: 3 - Alpha",
+  "Environment :: Console",
+  "License :: OSI Approved :: MIT License",
+  "Operating System :: OS Independent",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
+  "Topic :: Documentation",
+  "Topic :: Software Development :: Documentation",
+]
+
+[project.urls]
+Homepage = "https://github.com/sunniie/nexcanvas-drawio"
+Documentation = "https://github.com/sunniie/nexcanvas-drawio#readme"
+Issues = "https://github.com/sunniie/nexcanvas-drawio/issues"
+
+[project.scripts]
+nexcanvas = "nexcanvas.cli:main"
+
+[tool.setuptools]
+package-dir = { "" = "src" }
+
+[tool.setuptools.dynamic]
+version = { attr = "nexcanvas.__version__" }
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.setuptools.data-files]
+"share/nexcanvas/config" = ["config/*.json"]
+"share/nexcanvas/assets/catalog" = ["assets/catalog/*.json"]
+"share/nexcanvas/schemas" = ["schemas/*.json"]
+
