+++ src/nexcanvas/__init__.py
@@ -1,0 +1,4 @@
+"""NexCanvas Draw.io deterministic authoring toolkit."""
+
+# x-release-please-version
+__version__ = "0.1.0"
