+++ src/nexcanvas/__main__.py
@@ -1,0 +1,6 @@
+from .cli import main
+
+
+if __name__ == "__main__":
+    raise SystemExit(main())
+
