+++ src/nexcanvas/archetypes.py
