+++ src/nexcanvas/assets.py
