+++ src/nexcanvas/planning.py
