+++ src/nexcanvas/visual.py
